refactor(panel): retire AdminPanel, porting its tests onto ListPanel
All three panels render through ListPanel, and nothing imported the old
primitive any more except its own test file. AdminPanel.tsx and its stylesheet
go; app/components/AdminPanel/ stays for AdminPanelCollapseContext and
AdminPanelSeedContext, which are unrelated to the shell.

The test file is MOVED, not deleted. Its 14 tests cover behaviour ListPanel
still owns -- the opt-in collapse gate, the header action staying outside the
toggle button, the collapsed sliver being an ::after and not markup, the heading
semantics, aria-controls pointing at the body it hides -- and deleting them with
the component would have dropped the coverage while the behaviour stayed.

  tests/components/AdminPanel/AdminPanel.test.tsx   14 tests / 18 assertions
  tests/components/ListPanel/ListPanel.test.tsx      5 tests /  8 assertions (before)
  tests/components/ListPanel/ListPanel.test.tsx     20 tests / 30 assertions (after)

14 + 5 = 19 of those 20, and 18 + 8 = 26 of those 30, are the two files' tests
carried over unchanged. The only substantive edit was `action` becoming
`headerRight`. The extra test and its four assertions are new, below.

Added: the header and a row must present the same three sections in the same
order, which is what puts a header action and a row action on one rail. The
plan asked for this as a getComputedStyle comparison of grid-template-columns,
which would have been VACUOUS here -- next/jest stubs CSS modules, so neither
element has a computed grid and the assertion reads '' === '' and passes against
any stylesheet at all. Structure is what jsdom can actually see; the grid
declarations themselves are compared against the real SCSS in the following
commit.

The load-bearing-CSS docblock in AdminPanelRenderer is rewritten because it
named mechanisms that no longer exist: `flex: 1 1 220px` on a .rowMain that is
gone, and the ABSENCE of flex-wrap on a header that is no longer a flex row.
Those were a guarantee made of two things nobody had written down, either of
which a plausible edit restores. What protects width-independence now is
structural -- ListPanel's .row is a CSS grid with fixed tracks and a grid row
cannot wrap -- so the paragraph says that, and then keeps the three rules that
still carry weight, since a grid track cannot stop its own contents growing
taller: nowrap/ellipsis on every text slot, min-width: 0 on the 1fr track, and
no control exceeding its declared slot height.
